+++ NEWS.md
@@ -6,7 +6,7 @@ between bioregion 1.0.0 (CRAN release 2023-04-15) and the next CRAN release.
 * Modification of the way 'bioregion.pairwise.metric' object are controlled
 
 * Allow to (not) select 'formula metrics' in 
-similarity_dissimilarity_conversion()
+similarity_dissimilarity_conversion() with the new argument "include_formula"
 
 * Allow negative values in similarity() with the Euclidean metric